Fix bug with filenames that contain dots
Filenames that contain dots (for example `1.1-test.md`) were successfully inventoried in the TOC, but the files weren't crawled. This was due to the different mechanisms of generating IDs - the alias is extracted from the filename, while this helper function also filters out values, including dots. If the values don't match, the file is not added to the output.

If there are similiar bugs in the future, this should be the point to look at.